requirements.txt
python-dotenv
discord.py
aiohttp
SQLAlchemy
psycopg2
alembic